Uploading p5.js files from my laptop
I have programmed this mini application where I can experience a warm welcome sound clip at the start of the game with a text message guiding me to press enter key and continue, along with a graphic which expresses the user that this app is about a paint brush game.

After I press enter key, it shows me a blank yellow coloured canvas background. On this Canvas background I can toggle between 4 different brushes with a click of any key on the keyboard and draw on the canvas.

To conclude, its a small but quite fun little game with a lot of potential to be a better application in future.